--- lib/hipchat_exporter/root_path.rb
@@ -1,3 +1,0 @@
-module HipChatExporter
-  ROOT_PATH = File.expand_path('../..', __dir__)
-end
