+++ spec/room_spec.rb
@@ -1,13 +1,13 @@
-describe HipChatExporter::Room do
+describe Room do
   describe '.parse_csv' do
     it 'parse csv file and generate rooms' do
-      rooms = HipChatExporter::Room.parse_csv
-      expect(rooms[0].class).to eq HipChatExporter::Room
+      rooms = Room.parse_csv
+      expect(rooms[0].class).to eq Room
     end
   end
 
   describe '.generate' do
-    let(:room) { HipChatExporter::Room.generate(csv_row) }
+    let(:room) { Room.generate(csv_row) }
 
     context 'when name exists' do
       let(:csv_row) { [123, 'My Room'] }
@@ -26,14 +26,14 @@ describe HipChatExporter::Room do
 
   describe 'initialize' do
     context 'when name exists' do
-      let(:room) { HipChatExporter::Room.new(id: 123, name: 'My Room') }
+      let(:room) { Room.new(id: 123, name: 'My Room') }
 
       it { expect(room.id).to eq 123 }
       it { expect(room.name).to eq 'My Room' }
     end
 
     context 'when name not exist' do
-      let(:room) { HipChatExporter::Room.new(id: 123) }
+      let(:room) { Room.new(id: 123) }
 
       it { expect(room.id).to eq 123 }
       it { expect(room.name).to eq 'room_123' }
